refactors components and updates snapshots
