+++ src/components/reviews-list/reviews-list.jsx
@@ -7,15 +7,18 @@ const MAX_REVIEWS_TO_SHOW = 10;
 const ReviewsList = ({reviews}) => {
   const sortedReviews = reviews.slice(-MAX_REVIEWS_TO_SHOW).sort((a, b) => b.date - a.date);
 
-  return (
-    <ul className="reviews__list">
-      {sortedReviews.map((review) =>
-        <Review
-          review={review}
-          key={review.id}
-        />
-      )}
-    </ul>
+  return (reviews.length &&
+    <section className="property__reviews reviews">
+      <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">{reviews.length}</span></h2>
+      <ul className="reviews__list">
+        {sortedReviews.map((review) =>
+          <Review
+            review={review}
+            key={review.id}
+          />
+        )}
+      </ul>
+    </section>
   );
 };
